feat: #59 APIEndpoint 파일 분리 및 enum 추갸
